pan/genxml: Make sure we catch overflows on 32-bit integer fields
If we store the temporary value in a [u]int32_t, the result might be
silently truncated, making the overflow check in the u_pack helpers
useless.

Make sure we use 64-bit fields as soon as the size is greater than 31
bits to prevent that.

It forces us to fix some types in the xml and helper arguments (sint
wrongly defined as uint) and add explicit casts on u32 subtraction
whose result is stored in an s32.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/35555>

+++ src/gallium/drivers/panfrost/pan_jm.c
@@ -737,7 +737,7 @@ jm_emit_primitive(struct panfrost_batch *batch,
           * on Valhall, so we don't need to set that here
           */
       } else if (cfg.index_type) {
-         cfg.base_vertex_offset = draw->index_bias - ctx->offset_start;
+         cfg.base_vertex_offset = (int64_t)draw->index_bias - ctx->offset_start;
 
 #if PAN_ARCH <= 7
          cfg.indices = batch->indices;
